Add a pylint travis ci test
Add a test to the Travis CI build to check the program for style errors using pylint. Also change the build to use Python 3.4, as the curent version of pylint does not work with any newer versions of Python. In addition, add a configuration file for pylint to be used when testing the project.