Add has methods to check for tokens
Add several has methods to the JavaLikeScanner class in order to allow one to test for the existence of certain tokens, before getting them. Also restructure how the next methods work so that they make use of the has methods in order to be simpler.

Also add several tests of the new has methods of the JavaLikeScanner class.