Make various style improvements to the api
Move the first line of function docstrings to the second line, so that they are easier to read. Also fix the usage of return documentation, so that each is only on one line, as autodoc requires this setup to properly display the return info.

Also fix a small style issue.